feat: move country-of-concern banner to welcome screen and swap globe icon
Relocate the COC banner from above the chat input to inside the chat
empty-state, directly beneath the example-prompts welcome card. The
banner now appears only on the welcome screen, where it can sit as a
proper card under the introductory box (matching the spec mock).

Replace the 🌐 emoji with the tabler IconWorld outline glyph so the
title icon renders consistently across platforms.
